Update vimrc
I try to write good git commits, but I'm gonna be honest.

I don't remember what I changed when, only that it hasn't been commited
to the repo yet. So here we are. Hopefully it was something sensical;
I've noticed that there's a lot of comments added to the config, so
maybe that was what I did?

The diff history will tell the full story!