Update vimrc and xinitrc
My vimrc wasn't working properly for markdown files properly, though the
method I was trying to use to fix it properly was working for gitcommit
messages, so I'm not sure what's up with that.

The xinitrc change is to ensure that my custom colours for urxvt fire
properly, and as I intend to format my laptop soon I'll see if
everything plays well.