Update i3, i3blocks, vim and x files
I've been doing quite a lot of tinkering with my dotfiles. This is a
more decisive set of changes for my environment, with the most notable
changes being to i3 and i3blocks, altering the folder layout and
customization options. The latter now also includes a submodule for the
i3blocks-contrib repo, though I haven't played with it much.

There are also some minor changes to beets and sublime-related
configuration, while vim has been more thoroughly commented. In the
future I'm considering adding a README to each individual folder so they
can be easily read when browsing an online repo like github.