Update dunst and fish configuration files
I'm pretty happy with how Dunst is set-up for the time being - this
commit adds a small positioning change to it so it does not overlap with
my i3 tray. The fish-shell changes all relate to my plugin-setup, which
now includes fish-pipenv, automatically activating and using
pipenv-features when it changes into a relevant folder.