vim: Disable arrow keys, enable relative line numbers
I'm trying to become earnestly proficient in vim nowadays, and part of
that is forcing myself to navigate through things properly - to be
frank, I had no clue that relative line numbering was a thing, and now
that I have it enabled trying to navigate is much easier.

I've disabled my directional arrow keys to force myself to build the
muscle memory of navigating with hjkl; the first hurdle was jk and the
second will be hl, but what I really want to know is how people navigate
on a given line with numbers. I've been trying to eyeball the amount of
letters I'm jumping but that surely can't be how people do it, right?

Especially given that spacing and punctuation can count as words - I
haven't fully internalised the full list, so I find myself undershooting
all of the time.