feat: beets, KeePassXC, sway & zsh updates
Update the beets library and configuration: in the latter case,
it's simply formatting changed for the configuration file.

KeePassXC changes are somewhat superficial, as they are default
configuration values added to the configuration from triggering
their respective features on and off.

I have switched to sway (And wayland) from i3 and are slowly
creating a new zsh configuration to go with it. Previously, a
large amount of the system's programs and other desktop related
configuration options were all set through xorg files, so I'm
figuring out what needs to go where at the moment.